refactor: extract file iterator into multiple files
